+++ src/fable-library-py/fable_library/event.py
@@ -10,7 +10,8 @@ from .choice import (
 )
 from .observable import IObservable, IObserver, Observer
 from .option import Option, some, value
-from .util import IDisposable
+from .protocols import IDisposable
+from .util import DisposableBase
 
 
 _MISSING: Final[object] = object()
@@ -95,7 +96,7 @@ class Event[T](IEvent_2[T, T]):
         def dispose() -> None:
             self._removeHandler(callback)
 
-        return IDisposable.create(dispose)
+        return DisposableBase.create(dispose)
 
     def _addHandler(self, f: EventDelegate[T]) -> None:
         self.delegates.append(f)
@@ -200,7 +201,7 @@ class AnonymousEvent[U](IEvent_2[U, U]):
         def dispose() -> None:
             self._remove_handler(h)
 
-        return IDisposable.create(dispose)
+        return DisposableBase.create(dispose)
 
 
 def create_event[T](